Edit to remove the api module as I don't see a need to use it for this application. I want the single service. Write to use a better way of create the documentation. Write some use of swagger. It needs work, and I still don't understand Flash and how best to version. Not everything is working, but this is a better method right now of completing this POC and moving on to queueing. Publish because the endpoint works. The swagger file generates. And most seems good enough to write on.
